Village/tools/requirements.txt
# For image_sequence_to_bitmap.py
Pillow>=9.0.0
